// src/api/generated/model/report-dto.ts
/**
 * Generated by orval — do not edit manually.
 * MySweldo HRIS API 0.0.1
 */import type { ReportDtoFormat } from './report-dto-format';
import type { ReportDtoType } from './report-dto-type';

export interface ReportDto {
  id: string;
  type: ReportDtoType;
  format: ReportDtoFormat;
  fileName: string;
  periodStart?: string;
  periodEnd?: string;
  payrollRunId?: string;
  createdAt: string;
}